+++ src/c/config.c
@@ -2,5 +2,5 @@
 
 #include <cjson/config.h>
 
-size_t print_offset = 0;
-size_t default_hash_length = 4;
+uint64_t print_offset = 0;
+uint64_t default_hash_length = 4;
